requirements.txt
pytest
allure-pytest
pytest_bdd
selenium
jsonpickle
pony
pymysql
mysql-connector-python-rf
lxml-4.5.0-cp37-cp37m-win32.whl
